show upload in drawer
