+++ app/api/upload-url/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { S3Client } from '@aws-sdk/client-s3';
+import { S3Client, HeadObjectCommand } from '@aws-sdk/client-s3';
 import { S3RequestPresigner } from '@aws-sdk/s3-request-presigner';
 import { HttpRequest } from '@smithy/protocol-http';
 import { parseUrl } from '@smithy/url-parser';
@@ -34,6 +34,21 @@ export async function POST(req: NextRequest) {
     const bucketName = process.env.NEXT_PUBLIC_AWS_S3_BUCKET_NAME!;
     const key = filename;
 
+    // 🔍 Check if the file already exists in S3
+    let fileExists = false;
+    try {
+      await s3.send(new HeadObjectCommand({ Bucket: bucketName, Key: key }));
+      fileExists = true;
+    } catch (error: any) {
+      if (error.name !== 'NotFound') {
+        console.error('Error checking file existence:', error);
+        return NextResponse.json(
+          { error: 'Failed to check file existence' },
+          { status: 500 }
+        );
+      }
+    }
+
     const url = parseUrl(
       `https://${bucketName}.s3.${process.env.AWS_REGION}.amazonaws.com/${key}`
     );
@@ -51,6 +66,7 @@ export async function POST(req: NextRequest) {
     return NextResponse.json({
       uploadUrl: signedUrl,
       fileUrl: `https://${bucketName}.s3.${process.env.AWS_REGION}.amazonaws.com/${key}`,
+      fileExists,
     });
   } catch (error) {
     console.error('Error generating presigned URL:', error);
